requirements.txt
codecarbon==2.1.4
grpcio==1.53.0
numpy==1.23.4
Pillow==9.5.0
protobuf==3.20.2
torch>=2.0.0
torchvision>=0.15.1
tqdm==4.65.0
